# openapi.yaml
openapi: 3.1.0
info:
  title: Silicon Commit API
  version: 0.2.0
  description: Organization-scoped collaborative work for Carbons and Silicons. Contract 1 supports negotiation, private projects, retained history, isolated IAM sandboxes, email and optional telemetry.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
- url: https://commit.teamofsilicons.com/api/v1
security:
- bearerAuth: []
- oboAccess: []
  oboAppId: []
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: List projects
      operationId: listProjects
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/RequestId'
      - {name: status, in: query, schema: {$ref: '#/components/schemas/ProjectStatus'}}
      - {name: silicon_id, in: query, schema: {$ref: '#/components/schemas/ActorId'}}
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/CommitVersion'
      - $ref: '#/components/parameters/SupportedVersions'
      - $ref: '#/components/parameters/TestingAppSecret'
      - $ref: '#/components/parameters/TelemetryPreference'
      responses:
        '200':
          description: Organization-visible projects.
          headers:
            X-Request-ID: {$ref: '#/components/headers/XRequestId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProjectPage'}
        '429': {$ref: '#/components/responses/RateLimited'}
        default: {$ref: '#/components/responses/Error'}
    post:
      tags: [Projects]
      summary: Create a project
      operationId: createProject
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/CommitVersion'
      - $ref: '#/components/parameters/SupportedVersions'
      - $ref: '#/components/parameters/TestingAppSecret'
      - $ref: '#/components/parameters/TelemetryPreference'
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ProjectCreate'}}}}
      responses:
        '201':
          description: Project created.
          headers:
            X-Request-ID: {$ref: '#/components/headers/XRequestId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
            Idempotency-Replayed: {$ref: '#/components/headers/IdempotencyReplayed'}
            Location: {$ref: '#/components/headers/Location'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Project'}
        '429': {$ref: '#/components/responses/RateLimited'}
        default: {$ref: '#/components/responses/Error'}
  /projects/{project_id}:
    parameters:
    - $ref: '#/components/parameters/OrgId'
    - $ref: '#/components/parameters/RequestId'
    - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [Projects]
      summary: Get a project
      operationId: getProject
      responses:
        '200':
          description: Project.
          headers:
            X-Request-ID: {$ref: '#/components/headers/XRequestId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Project'}
        '429': {$ref: '#/components/responses/RateLimited'}
        default: {$ref: '#/components/responses/Error'}
      parameters:
      - $ref: '#/components/parameters/CommitVersion'
      - $ref: '#/components/parameters/SupportedVersions'
      - $ref: '#/components/parameters/TestingAppSecret'
      - $ref: '#/components/parameters/TelemetryPreference'
    patch:
      tags: [Projects]
      summary: Update a project
      operationId: updateProject
      parameters: [$ref: '#/components/parameters/IdempotencyKey', $ref: '#/components/parameters/CommitVersion', $ref: '#/components/parameters/SupportedVersions', $ref: '#/components/parameters/TestingAppSecret', $ref: '#/components/parameters/TelemetryPreference']
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, minProperties: 1, additionalProperties: false, properties: {name: {$ref: '#/components/schemas/ProjectNameText'}, status: {$ref: '#/components/schemas/ProjectPatchStatus'}, silicon_ids: {
                  type: array, maxItems: 100, uniqueItems: true, description: 'Complete replacement set; it must retain the project creator.', items: {$ref: '#/components/schemas/ActorId'}}, description: {type: string,
                    maxLength: 100000, default: ''}, attachments: {type: array, maxItems: 20, items: {$ref: '#/components/schemas/AttachmentUrl'}}, private: {type: boolean, default: false}, carbon_ids: {type: array,
                    maxItems: 100, uniqueItems: true, items: {$ref: '#/components/schemas/ActorId'}}, tags: {type: array, maxItems: 100, items: {type: string, maxLength: 255}}}}
      responses:
        '200':
          description: Updated project.
          headers:
            X-Request-ID: {$ref: '#/components/headers/XRequestId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
            Idempotency-Replayed: {$ref: '#/components/headers/IdempotencyReplayed'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Project'}
        '429': {$ref: '#/components/responses/RateLimited'}
        default: {$ref: '#/components/responses/Error'}
  /projects/{project_id}/diary:
    get:
      tags: [Project Diary]
      summary: Get the project diary
      operationId: getProjectDiary
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/CommitVersion'
      - $ref: '#/components/parameters/SupportedVersions'
      - $ref: '#/components/parameters/TestingAppSecret'
      - $ref: '#/components/parameters/TelemetryPreference'
      responses:
        '200':
          description: Markdown diary.
          headers:
            X-Request-ID: {$ref: '#/components/headers/XRequestId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
            ETag: {$ref: '#/components/headers/ETag'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Diary'}
        '429': {$ref: '#/components/responses/RateLimited'}
        default: {$ref: '#/components/responses/Error'}
    put:
      tags: [Project Diary]
      summary: Replace the project diary
      operationId: updateProjectDiary
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/IfMatch'
      - $ref: '#/components/parameters/CommitVersion'
      - $ref: '#/components/parameters/SupportedVersions'
      - $ref: '#/components/parameters/TestingAppSecret'
      - $ref: '#/components/parameters/TelemetryPreference'
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [markdown], additionalProperties: false, properties: {markdown: {$ref: '#/components/schemas/DiaryMarkdown'}}}}}}
      responses:
        '200':
          description: Updated diary.
          headers:
            X-Request-ID: {$ref: '#/components/headers/XRequestId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
            ETag: {$ref: '#/components/headers/ETag'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Diary'}
        '409': {$ref: '#/components/responses/Error'}
        '429': {$ref: '#/components/responses/RateLimited'}
        default: {$ref: '#/components/responses/Error'}
  /projects/{project_id}/entries:
    get:
      tags: [Project Work]
      summary: List project activity
      operationId: listProjectEntries
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/CommitVersion'
      - $ref: '#/components/parameters/SupportedVersions'
      - $ref: '#/components/parameters/TestingAppSecret'
      - $ref: '#/components/parameters/TelemetryPreference'
      responses:
        '200':
          description: Blockers, updates, and completion entries in reverse creation order.
          headers:
            X-Request-ID: {$ref: '#/components/headers/XRequestId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProjectEntryPage'}
        '429': {$ref: '#/components/responses/RateLimited'}
        default: {$ref: '#/components/responses/Error'}
  /projects/{project_id}/tasks:
    get:
      tags: [Project Work]
      summary: List project tasks
      operationId: listProjectTasks
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/CommitVersion'
      - $ref: '#/components/parameters/SupportedVersions'
      - $ref: '#/components/parameters/TestingAppSecret'
      - $ref: '#/components/parameters/TelemetryPreference'
      responses:
        '200':
          description: Tasks and nested subtasks in reverse creation order.
          headers:
            X-Request-ID: {$ref: '#/components/headers/XRequestId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProjectTaskPage'}
        '429': {$ref: '#/components/responses/RateLimited'}
        default: {$ref: '#/components/responses/Error'}
    post:
      tags: [Project Work]
      summary: Create a project task
      operationId: createProjectTask
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/CommitVersion'
      - $ref: '#/components/parameters/SupportedVersions'
      - $ref: '#/components/parameters/TestingAppSecret'
      - $ref: '#/components/parameters/TelemetryPreference'
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ProjectTaskCreate'}}}}
      responses:
        '201':
          description: Task or subtask created.
          headers:
            X-Request-ID: {$ref: '#/components/headers/XRequestId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
            Idempotency-Replayed: {$ref: '#/components/headers/IdempotencyReplayed'}
            Location: {$ref: '#/components/headers/Location'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProjectTask'}
        '429': {$ref: '#/components/responses/RateLimited'}
        default: {$ref: '#/components/responses/Error'}
  /projects/{project_id}/tasks/{task_id}:
    patch:
      tags: [Project Work]
      summary: Update a project task
      operationId: updateProjectTask
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/RequestId'
      - &id001
        $ref: '#/components/parameters/ProjectId'
      - &id002 {name: task_id, in: path, required: true, schema: {type: string, format: uuid}}
      - $ref: '#/components/parameters/CommitVersion'
      - $ref: '#/components/parameters/SupportedVersions'
      - $ref: '#/components/parameters/TestingAppSecret'
      - $ref: '#/components/parameters/TelemetryPreference'
      requestBody: {required: true, content: {application/json: {schema: {type: object, minProperties: 1, additionalProperties: false, properties: {title: {$ref: '#/components/schemas/TitleText'}, description: {$ref: '#/components/schemas/DescriptionText'},
                status: {$ref: '#/components/schemas/TodoStatus'}, assigned_to: {oneOf: [$ref: '#/components/schemas/ActorId', type: 'null']}}}}}}
      responses:
        '200':
          description: Updated task.
          headers:
            X-Request-ID: {$ref: '#/components/headers/XRequestId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProjectTask'}
        '429': {$ref: '#/components/responses/RateLimited'}
        default: {$ref: '#/components/responses/Error'}
    delete:
      tags:
      - Projects
      operationId: deleteProjectTask
      summary: Remove task, descendants and linked todos
      responses:
        '204':
          description: Remove task, descendants and linked todos
        default:
          $ref: '#/components/responses/Error'
      parameters:
      - $ref: '#/components/parameters/CommitVersion'
      - $ref: '#/components/parameters/SupportedVersions'
      - $ref: '#/components/parameters/TestingAppSecret'
      - $ref: '#/components/parameters/TelemetryPreference'
      - *id001
      - *id002
  /projects/{project_id}/blockers:
    post:
      tags: [Project Work]
      summary: Add a project blocker
      operationId: createBlocker
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/CommitVersion'
      - $ref: '#/components/parameters/SupportedVersions'
      - $ref: '#/components/parameters/TestingAppSecret'
      - $ref: '#/components/parameters/TelemetryPreference'
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/BlockerCreate'}}}}
      responses:
        '201':
          description: Blocker created.
          headers:
            X-Request-ID: {$ref: '#/components/headers/XRequestId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
            Idempotency-Replayed: {$ref: '#/components/headers/IdempotencyReplayed'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProjectEntry'}
        '429': {$ref: '#/components/responses/RateLimited'}
        default: {$ref: '#/components/responses/Error'}
  /projects/{project_id}/updates:
    post:
      tags: [Project Work]
      summary: Add a project update
      operationId: createProjectUpdate
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/CommitVersion'
      - $ref: '#/components/parameters/SupportedVersions'
      - $ref: '#/components/parameters/TestingAppSecret'
      - $ref: '#/components/parameters/TelemetryPreference'
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ProjectEntryCreate'}}}}
      responses:
        '201':
          description: Update created.
          headers:
            X-Request-ID: {$ref: '#/components/headers/XRequestId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
            Idempotency-Replayed: {$ref: '#/components/headers/IdempotencyReplayed'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProjectEntry'}
        '429': {$ref: '#/components/responses/RateLimited'}
        default: {$ref: '#/components/responses/Error'}
  /projects/{project_id}/completion:
    post:
      tags: [Project Work]
      summary: Complete a project
      operationId: completeProject
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/CommitVersion'
      - $ref: '#/components/parameters/SupportedVersions'
      - $ref: '#/components/parameters/TestingAppSecret'
      - $ref: '#/components/parameters/TelemetryPreference'
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ProjectEntryCreate'}}}}
      responses:
        '201':
          description: Completion entry created and project completed.
          headers:
            X-Request-ID: {$ref: '#/components/headers/XRequestId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
            Idempotency-Replayed: {$ref: '#/components/headers/IdempotencyReplayed'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProjectEntry'}
        '429': {$ref: '#/components/responses/RateLimited'}
        default: {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: IAM access token}
    oboAccess: {type: apiKey, in: header, name: X-IAM-OBO-Access-Proof, description: 'Short-lived, single-use IAM OBO Access proof.'}
    oboAppId: {type: apiKey, in: header, name: X-App-ID, description: Issuer application ID paired with the OBO proof.}
  parameters:
    OrgId: {name: X-Org-ID, in: header, required: true, description: Public organization context which must match current IAM authority., schema: {$ref: '#/components/schemas/OrganizationId'}}
    RequestId: {name: X-Request-ID, in: header, required: false, description: Optional visible-ASCII correlation identifier which is returned unchanged when valid., schema: {type: string, minLength: 1, maxLength: 128,
          pattern: '^[!-~]+$'}}
...
    ProjectId: {name: project_id, in: path, required: true, description: 'Project UUID or exact stable UID, bounded to 2,048 UTF-8 bytes. Bare slugs are not accepted; percent-encode a UID as exactly one path segment.',
        schema: {type: string, maxLength: 2048}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, description: Opaque case-sensitive replay key containing visible ASCII only., schema: {type: string, minLength: 8, maxLength: 255, pattern: '^[!-~]+$'}}
...
    Cursor: {name: cursor, in: query, schema: {type: string, minLength: 1, maxLength: 1024}}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
    CommitVersion:
      name: X-Commit-API-Version
      in: header
      required: false
      description: Explicit API contract; currently 1.
      schema:
        type: string
    SupportedVersions:
      name: X-Commit-Supported-Versions
      in: header
      required: false
      description: Comma-separated supported versions; currently must include 1.
      schema:
        type: string
    TestingAppSecret:
      name: X-Testing-App-Secret
      in: header
      required: false
      description: Imported IAM application secret, never a URL parameter. Selects an isolated sandbox. Do not combine with X-Testing-Environment-Key.
      schema:
        type: string
    TelemetryPreference:
      name: X-Commit-Telemetry
      in: header
      required: false
      description: Set off to disable diagnostics for this request.
      schema:
        type: string
  responses:
    Error:
      description: Stable redacted error envelope.
      headers:
        X-Request-ID: {$ref: '#/components/headers/XRequestId'}
        Cache-Control: {$ref: '#/components/headers/CacheControl'}
        WWW-Authenticate: {$ref: '#/components/headers/WwwAuthenticate'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    RateLimited:
      description: A dependency or request policy temporarily rate-limited the operation.
      headers:
        X-Request-ID: {$ref: '#/components/headers/XRequestId'}
        Cache-Control: {$ref: '#/components/headers/CacheControl'}
        Retry-After: {$ref: '#/components/headers/RetryAfter'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
  headers:
    XRequestId: {description: Request correlation identifier., schema: {type: string, minLength: 1, maxLength: 128, pattern: '^[!-~]+$'}}
    CacheControl: {description: Responses are not cacheable by shared or private caches., schema: {type: string, const: no-store}}
    IdempotencyReplayed: {description: Whether this exact response was replayed from durable idempotency state., schema: {type: boolean}}
    Location: {description: Absolute canonical URL of the created resource., schema: {type: string, format: uri}}
...
  schemas:
...
    ActorId:
      type: string
      minLength: 1
      maxLength: 255
      pattern: '^(?=[\s\S]*\S)[^\u0000-\u001F\u007F-\u009F]*$'
      description: Stable nonblank public IAM Carbon or Silicon identifier. Raw surrounding whitespace counts toward the limit and is trimmed; control characters are forbidden.
...
    ProjectStatus: {type: string, enum: [completed, blocked, canceled, in_progress, yet_to_start]}
...
    ProjectCreate:
      type: object
      required:
      - name
      additionalProperties: false
      properties: {name: {$ref: '#/components/schemas/ProjectNameText'}, silicon_ids: {type: array, maxItems: 100, uniqueItems: true, items: {$ref: '#/components/schemas/ActorId'}}, description: {type: string, maxLength:
            100000, default: ''}, attachments: {type: array, maxItems: 20, items: {$ref: '#/components/schemas/AttachmentUrl'}}, private: {type: boolean, default: false}, carbon_ids: {type: array, maxItems: 100,
            uniqueItems: true, items: {$ref: '#/components/schemas/ActorId'}}, tags: {type: array, maxItems: 100, items: {type: string, maxLength: 255}}, tasks: {type: array, maxItems: 1000, items: {$ref: '#/components/schemas/ProjectSeedTask'}}}
...
    ProjectPage:
      type: object
      required: [items, next_cursor]
      additionalProperties: false
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/Project'}}
        next_cursor: {type: [string, 'null'], maxLength: 1024}
...